% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{bar_plot}
\alias{bar_plot}
\title{Create a bar plot}
\usage{
bar_plot(
  subset_df,
  var,
  colsplit_by = "No split",
  rowsplit_by = "No split",
  colorsplit_by = "No split",
  plot_height = 470
)
}
\arguments{
\item{var}{An object of class "string", name of the variable to plot (must be categorical)}

\item{colsplit_by}{(optional) An object of class "string", name of the variable to use in facetgrid col (must be categorical)}

\item{rowsplit_by}{(optional) An object of class "string", name of the variable to use in facetgrid row (must be categorical)}

\item{colorsplit_by}{(optional) An object of class "string", name of the variable to spli by color (must be categorical)}

\item{dataset}{An object of class "data.frame", the dataset to plot}
}
\value{
Returns an object of class "plotly". the function create a bar plot based on input
}
\description{
Create a bar plot using ggplot and plotly for a given variable
and possibility to split on row, col and color.
}
\examples{
bar_plot(dataset, "workex", colsplit_by = "gender")

}
