% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{distribution_plot}
\alias{distribution_plot}
\title{Create a distribution plot}
\usage{
distribution_plot(
  dataset,
  var,
  type = "count",
  colsplit_by = "No split",
  rowsplit_by = "No split",
  colorsplit_by = "No split",
  plot_height = 470
)
}
\arguments{
\item{dataset}{An object of class "data.frame", the dataset to plot}

\item{var}{An object of class "string", name of the variable to plot (must be numerical)}

\item{type}{An object of class "string", plot type ("count", "density")}

\item{colsplit_by}{(optional) An object of class "string", name of the variable to use in facetgrid col (must be categorical)}

\item{rowsplit_by}{(optional) An object of class "string", name of the variable to use in facetgrid row (must be categorical)}

\item{colorsplit_by}{(optional) An object of class "string", name of the variable to spli by color (must be categorical)}
}
\value{
Returns an object of class "plotly". the function create a distribution plot based on input
}
\description{
Create a distribution plot using ggplot and plotly for a given var
and possibility to split on row, col and color.
Median and mean are automatically added as vertical lines
}
\examples{
distribution_plot(dataset, "hsc_p","count",rowsplit_by = "gender")

}
