% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{employ_rate}
\alias{employ_rate}
\title{compute the employment rate}
\usage{
employ_rate(dataset, split_by, value)
}
\arguments{
\item{dataset}{An object of class "data.frame", the dataset to plot}

\item{split_by}{An object of class "string", name of the variable to group_by before computing (must be categorical)}

\item{value}{An object of class "string", to filter on a category in the split_by column}
}
\value{
Returns an object of class "string", with the employment rate in it
}
\description{
compute the employment rate for the category given by value in split_by
It has an option 'no_split' in case the used want a general profil
}
\examples{
employ_rate(dataset, "gender", "M")
}
