% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{gen_sankey_data}
\alias{gen_sankey_data}
\title{generate the data use by the sankey plot}
\usage{
gen_sankey_data(data, where_var, to_var, where_to_id, existing_data = NULL)
}
\arguments{
\item{data}{An object of class "data.frame", the dataset to plot}

\item{where_var}{An object of class "string", variable to place on the left side of the sankey plot}

\item{to_var}{An object of class "string",  variable to place on the right side of the sankey plot}

\item{where_to_id}{An object of class "data.frame", the matching table from gen_sankey_id}

\item{existing_data}{An object of class "data.frame", generated by a previous call of this function to be stacked with new data}
}
\value{
Returns an object of class "dataframe", the data to be used in sankey plot
}
\description{
generate id for each categories used in the sankey plot, id start from 0 to n-1 categories
the function can take a data.frame containing the data for the sankey plot
}
\examples{
gen_sankey_data(dataset, "hsc_b", "status", where_to_id, existing_data)
}
