% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{gen_sankey_id}
\alias{gen_sankey_id}
\title{generate id for each categories}
\usage{
gen_sankey_id(data, where_var, to_var, existing_ids = data.frame(name = c()))
}
\arguments{
\item{data}{An object of class "data.frame", the dataset to plot}

\item{where_var}{An object of class "string", variable to place on the left side of the sankey plot}

\item{to_var}{An object of class "string", variable to place on the right side of the sankey plot}

\item{existing_ids}{An object of class "data.frame", containing existing id in order to not add them twice}
}
\value{
Returns an object of class "dataframe", a matching table between
categories in where_var, to_var and the given ids
}
\description{
generate id for each categories used in the sankey plot, id start from 0 to n-1 categories
the function can take a data.frame containing existing id in order to not add them twice
}
\examples{
gen_sankey_id(dataset, "hsc_b", "status", existing_ids)
}
