% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{min_salary}
\alias{min_salary}
\title{compute the min salary}
\usage{
min_salary(dataset, split_by, value)
}
\arguments{
\item{dataset}{An object of class "data.frame", the dataset to plot}

\item{split_by}{An object of class "string", name of the variable to group_by before computing (must be categorical)}

\item{value}{to An object of class "string", filter on a category in the split_by column}
}
\value{
Returns an object of class "string", with the min salary in it
}
\description{
compute the min salary for the category given by value in split_by
It has an option 'no_split' in case the used want a general profil
}
\examples{
min_salary(dataset, "gender", "M")
}
