% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{plot_sankey}
\alias{plot_sankey}
\title{generate the sankey plot}
\usage{
plot_sankey(sankey_data, sankey_ids, plot_height = "470px")
}
\arguments{
\item{sankey_data}{An object of class "data.frame", the dataset to plot}

\item{sankey_ids}{An object of class "data.frame", the matching table between categories and ids}
}
\value{
Returns an object of class "plotly", the sankey plot
}
\description{
generate the sankey plot based on the data and the ids generated by gen_sankey_data
and gen_sankey_ids
}
\examples{
plot_sankey(sankey_data, sankey_ids)
}
