% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{plot_sankey_recursive}
\alias{plot_sankey_recursive}
\title{generate the sankey plot on multiple steps}
\usage{
plot_sankey_recursive(dataset, stages)
}
\arguments{
\item{dataset}{An object of class "data.frame", the dataset to plot}

\item{stages}{An object of class "vector", all the stage to use (where and to variables)}
}
\value{
Returns an object of class "plotly", the sankey plot for multiple stages
}
\description{
combine gen_sankey_ids, gen_sankey_data and plot_sankey to compute a sankey plot
on multiple stages
}
\examples{
plot_sankey_recursive(dataset, c("hsc_b", "workex", "status"))
}
