% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{prepare_radar_data}
\alias{prepare_radar_data}
\title{Prepare the data for radar plot}
\usage{
prepare_radar_data(dataset, splitby)
}
\arguments{
\item{dataset}{An object of class "data.frame", the dataset to plot}

\item{split_by}{An object of class "string", name of the variable to group by the data before computing the median (must be categorical)}
}
\value{
Returns an object of class "data.frame". Median for each group and all numerical variables
}
\description{
Prepare the data for radar plot by selecting the right columns
grouping by the split_by var and computing the median
It has an option 'no_split' in case the used want a generaa profil
}
\examples{
prepare_radar_data(dataset, "gender")
}
