% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Create a scatter plot}
\usage{
scatter_plot(
  subset_df,
  var_x,
  var_y,
  colsplit_by = "No split",
  rowsplit_by = "No split",
  colorsplit_by = "No split",
  plot_height = 470
)
}
\arguments{
\item{var_x}{An object of class "string", name of the variable to plot on x axis (must be numerical)}

\item{var_y}{An object of class "string", name of the variable to plot on y axis (must be numerical)}

\item{colsplit_by}{(optional) An object of class "string", name of the variable to use in facetgrid col (must be categorical)}

\item{rowsplit_by}{(optional) An object of class "string", name of the variable to use in facetgrid row (must be categorical)}

\item{colorsplit_by}{(optional) An object of class "string",  name of the variable to spli by color (must be categorical)}

\item{dataset}{An object of class "data.frame", the dataset to plot}
}
\value{
Returns an object of class "plotly". the function create a scatter plot based on input
}
\description{
Create a scatter plot using ggplot and plotly for x and y given variables and possibility to split on row, col and color
}
\examples{
scatter_plot(dataset, "hsc_s", "salary", colorsplit_by = "gender")

}
